+++ primitives/src/lib.rs
@@ -42,7 +42,11 @@ pub use units::*;
 #[cfg(feature = "alloc")]
 pub use self::locktime::{absolute, relative};
 #[doc(inline)]
-pub use self::{pow::CompactTarget, sequence::Sequence};
+pub use self::{
+    pow::CompactTarget,
+    sequence::Sequence,
+    transaction::{Txid, Wtxid},
+};
 
 #[rustfmt::skip]
 #[allow(unused_imports)]
